💄 style: improve thinking style (#5602)
* refactor the thinking comp

* update locale
