+++ locales/bg-BG/chat.json
@@ -82,10 +82,6 @@
       }
     }
   },
-  "reasoning": {
-    "thinking": "В дълбочинно размисъл",
-    "thought": "Дълбоко помислих (отне ми {{duration}} секунди)"
-  },
   "regenerate": "Прегенерирай",
   "roleAndArchive": "Роля и архив",
   "searchAgentPlaceholder": "Търсач на помощ...",
