+++ locales/bg-BG/components.json
@@ -85,5 +85,10 @@
   "ModelSwitchPanel": {
     "emptyModel": "Няма активирани модели, моля, посетете настройките и ги активирайте",
     "provider": "Доставчик"
+  },
+  "Thinking": {
+    "thinking": "В процес на дълбочинно размисъл...",
+    "thought": "Дълбочинно размислено",
+    "thoughtWithDuration": "Дълбочинно размислено (време: {{duration}} секунди)"
   }
 }
