+++ locales/de-DE/chat.json
@@ -82,10 +82,6 @@
       }
     }
   },
-  "reasoning": {
-    "thinking": "Tiefes Nachdenken",
-    "thought": "Tiefgründig nachgedacht (Dauer: {{duration}} Sekunden)"
-  },
   "regenerate": "Neu generieren",
   "roleAndArchive": "Rolle und Archiv",
   "searchAgentPlaceholder": "Suchassistent...",
