+++ locales/es-ES/chat.json
@@ -82,10 +82,6 @@
       }
     }
   },
-  "reasoning": {
-    "thinking": "Pensando profundamente",
-    "thought": "He reflexionado profundamente (durante {{duration}} segundos)"
-  },
   "regenerate": "Regenerar",
   "roleAndArchive": "Rol y archivo",
   "searchAgentPlaceholder": "Asistente de búsqueda...",
