+++ locales/ja-JP/components.json
@@ -85,5 +85,10 @@
   "ModelSwitchPanel": {
     "emptyModel": "有効なモデルがありません。設定に移動して有効にしてください。",
     "provider": "プロバイダー"
+  },
+  "Thinking": {
+    "thinking": "深く考えています...",
+    "thought": "深く考えました",
+    "thoughtWithDuration": "深く考えました（所要時間 {{duration}} 秒）"
   }
 }
