+++ locales/nl-NL/chat.json
@@ -82,10 +82,6 @@
       }
     }
   },
-  "reasoning": {
-    "thinking": "Diep aan het denken",
-    "thought": "Diep nagedacht (tijd: {{duration}} seconden)"
-  },
   "regenerate": "Opnieuw genereren",
   "roleAndArchive": "Rol en archief",
   "searchAgentPlaceholder": "Zoekassistent...",
