+++ locales/nl-NL/components.json
@@ -85,5 +85,10 @@
   "ModelSwitchPanel": {
     "emptyModel": "No enabled model, please go to settings to enable.",
     "provider": "Provider"
+  },
+  "Thinking": {
+    "thinking": "Diep in gedachten...",
+    "thought": "Diep nagedacht",
+    "thoughtWithDuration": "Diep nagedacht (geduurd {{duration}} seconden)"
   }
 }
