+++ locales/pl-PL/chat.json
@@ -82,10 +82,6 @@
       }
     }
   },
-  "reasoning": {
-    "thinking": "Głęboko myślę",
-    "thought": "Głęboko myślałem (czas: {{duration}} sekund)"
-  },
   "regenerate": "Wygeneruj ponownie",
   "roleAndArchive": "Rola i archiwum",
   "searchAgentPlaceholder": "Wyszukaj pomocnika...",
