+++ locales/pl-PL/components.json
@@ -85,5 +85,10 @@
   "ModelSwitchPanel": {
     "emptyModel": "Brak włączonych modeli, przejdź do ustawień i włącz je",
     "provider": "Dostawca"
+  },
+  "Thinking": {
+    "thinking": "Głęboko myślę...",
+    "thought": "Już głęboko myślałem",
+    "thoughtWithDuration": "Już głęboko myślałem (czas: {{duration}} sekund)"
   }
 }
