+++ locales/pt-BR/chat.json
@@ -82,10 +82,6 @@
       }
     }
   },
-  "reasoning": {
-    "thinking": "Pensando profundamente",
-    "thought": "Pensou profundamente (tempo gasto: {{duration}} segundos)"
-  },
   "regenerate": "Regenerar",
   "roleAndArchive": "Função e Arquivo",
   "searchAgentPlaceholder": "Assistente de busca...",
