+++ locales/ru-RU/chat.json
@@ -82,10 +82,6 @@
       }
     }
   },
-  "reasoning": {
-    "thinking": "Глубокое размышление",
-    "thought": "Глубоко размышлял (время: {{duration}} секунд)"
-  },
   "regenerate": "Сгенерировать заново",
   "roleAndArchive": "Роль и архив",
   "searchAgentPlaceholder": "Поиск помощника...",
