+++ locales/ru-RU/components.json
@@ -85,5 +85,10 @@
   "ModelSwitchPanel": {
     "emptyModel": "Нет активированных моделей. Пожалуйста, перейдите в настройки и включите модель",
     "provider": "Поставщик"
+  },
+  "Thinking": {
+    "thinking": "Глубокое размышление...",
+    "thought": "Глубоко размышлено",
+    "thoughtWithDuration": "Глубоко размышлено (время: {{duration}} секунд)"
   }
 }
