+++ locales/zh-TW/components.json
@@ -85,5 +85,10 @@
   "ModelSwitchPanel": {
     "emptyModel": "沒有啟用的模型，請前往設定開啟",
     "provider": "提供商"
+  },
+  "Thinking": {
+    "thinking": "深度思考中...",
+    "thought": "已深度思考",
+    "thoughtWithDuration": "已深度思考（用時 {{duration}} 秒）"
   }
 }
